feat(utils): add random function and create types folder
- Add generic random function
- Create types folder for shared types
- Add typings for random function
