+++ src/assets/images/answer/scissors.png
