+++ src/assets/images/choices/paper.png
