+++ src/assets/images/choices/rock.png
