+++ src/assets/images/choices/scissors.png
