+++ src/core/computer.ts
@@ -1,0 +1,9 @@
+/**
+ *  Choix aléatoire du Computer
+ */
+import { randomElement } from "../shared/utils/random";
+const choices = ["rock", "paper", "scissors"];
+
+const getComputerChoice = (): string => randomElement(choices);
+
+export { getComputerChoice, choices };
