+++ src/shared/types/choices.ts
@@ -1,0 +1,1 @@
+export type Choices = "rock" | "paper" | "scissors";
