+++ src/shared/types/player-config.ts
@@ -1,0 +1,8 @@
+export interface PlayerConfig {
+  name: string;
+  score: number;
+  choice: string;
+  win: boolean;
+  lose: boolean;
+  draw: boolean;
+}
