+++ src/shared/utils/random.ts
@@ -1,0 +1,2 @@
+export const randomElement = <A>(array: A[]): A =>
+  array[~~(Math.random() * array.length)];
